# packages/api-client/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ecommerce Platform API",
    "version": "1.0.0",
    "description": "REST API for the ecommerce platform catalog, auth, cart, orders, health, and uploads."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local dev"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token from POST /api/auth/login or /api/auth/refresh"
      }
    },
...
  },
  "paths": {
...
    "/api/admin/users": {
      "get": {
        "tags": ["Admin"],
        "operationId": "listAdminUsers",
        "summary": "List users (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "status",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "search",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": ["admin", "customer"]
            },
            "required": false,
            "name": "role",
            "in": "query"
          },
          {
            "schema": {
              "type": "boolean",
              "nullable": true
            },
            "required": false,
            "name": "isActive",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "email": {
                            "type": "string",
                            "format": "email"
                          },
                          "role": {
                            "type": "string",
                            "enum": ["admin", "customer"]
                          },
                          "phone": {
                            "type": "string"
                          },
                          "emailVerified": {
                            "type": "boolean"
                          },
                          "isActive": {
                            "type": "boolean"
                          },
                          "orderCount": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "name",
                          "email",
                          "role",
                          "emailVerified",
                          "isActive",
                          "orderCount",
                          "createdAt"
                        ]
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    }
                  },
                  "required": ["data", "total", "page", "limit"]
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users/{id}/status": {
      "patch": {
        "tags": ["Admin"],
        "operationId": "updateAdminUserStatus",
        "summary": "Enable or disable a user (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Resource ID"
            },
            "required": true,
            "description": "Resource ID",
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "isActive": {
                    "type": "boolean"
                  }
                },
                "required": ["isActive"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string",
                      "format": "email"
                    },
                    "role": {
                      "type": "string",
                      "enum": ["admin", "customer"]
                    },
                    "phone": {
                      "type": "string"
                    },
                    "emailVerified": {
                      "type": "boolean"
                    },
                    "isActive": {
                      "type": "boolean"
                    },
                    "orderCount": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "email",
                    "role",
                    "emailVerified",
                    "isActive",
                    "orderCount",
                    "createdAt"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}